+++ src/models/wallet.rs
@@ -8,7 +8,7 @@ pub struct WalletRow {
     pub nonce: u64,
     #[serde(rename = "type")]
     pub wallet_type: String,
-    pub person_id: u16
+    pub person_id: u64
 }
 
 #[allow(dead_code)]
